Added splits sync task
