+++ src/main/java/io/split/android/client/service/SplitTask.java
@@ -1,4 +1,4 @@
-package io.split.android.client.backend;
+package io.split.android.client.service;
 
 public interface SplitTask {
     void execute();
