+++ src/main/java/io/split/android/client/service/SplitsSyncTask.java
@@ -1,14 +1,18 @@
-package io.split.android.client.backend;
+package io.split.android.client.service;
 
-import io.split.android.client.backend.SplitTask;
-import io.split.android.client.backend.splits.HttpSplitFetcher;
-import io.split.android.client.backend.splits.SplitFetcherV2;
+import io.split.android.client.dtos.SplitChange;
+import io.split.android.client.service.SplitTask;
+import io.split.android.client.service.splits.HttpSplitFetcher;
+import io.split.android.client.service.splits.SplitChangeProcessor;
+import io.split.android.client.service.splits.SplitFetcherV2;
 import io.split.android.client.storage.splits.SplitsStorage;
+import io.split.android.client.utils.Logger;
 
 public class SplitsSyncTask implements SplitTask {
 
     SplitFetcherV2 mSplitFetcher;
     SplitsStorage mSplitsStorage;
+    SplitChangeProcessor mSplitChangeProcessor;
 
     public SplitsSyncTask(SplitFetcherV2 splitFetcher, SplitsStorage splitsStorage) {
         mSplitFetcher = splitFetcher;
@@ -17,6 +21,12 @@ public class SplitsSyncTask implements SplitTask {
 
     @Override
     public void execute() {
-
+        try {
+            // TODO: Ask tincho why mSplitsStorage.getTill()
+            SplitChange splitChange = mSplitFetcher.execute(mSplitsStorage.getTill());
+            mSplitsStorage.update(mSplitChangeProcessor.process(splitChange));
+        } catch (IllegalStateException e) {
+            Logger.e("Error while executing splits sync task: " + e.getLocalizedMessage());
+        }
     }
 }
